Removed binary_ops.rs
