--- src/tensor/binary_ops.rs
@@ -1,39 +1,0 @@
-use crate::dtype::RawDataType;
-use crate::tensor::flags::TensorFlags;
-use crate::{Tensor};
-use std::ops::{AddAssign, BitAndAssign, BitOrAssign, DivAssign, MulAssign, RemAssign, ShlAssign,  ShrAssign, SubAssign};
-
-
-macro_rules! define_binary_iop {
-    ( $trait_: ident, $operator: tt, $method: ident ) => {
-    impl<T: RawDataType + $trait_> $trait_<Tensor<'_, T>> for Tensor<'_, T> {
-        fn $method(&mut self, rhs: Tensor<'_, T>) {
-            *self $operator &rhs
-        }
-    }
-
-    impl<T: RawDataType + $trait_> $trait_<&Tensor<'_, T>> for Tensor<'_, T> {
-        fn $method(&mut self, rhs: &Tensor<'_, T>) {
-            if !self.flags.contains(TensorFlags::Writeable) {
-                panic!("Tensor is readonly");
-            }
-
-            let rhs = rhs.broadcast_to(&self.shape);
-
-            for (lhs, rhs) in self.flatiter_ptr().zip(rhs.flatiter()) {
-                unsafe { *lhs $operator rhs; }
-            }
-        }
-    }
-    };
-}
-
-define_binary_iop!(AddAssign, +=, add_assign);
-define_binary_iop!(SubAssign, -=, sub_assign);
-define_binary_iop!(MulAssign, *=, mul_assign);
-define_binary_iop!(DivAssign, /=, div_assign);
-define_binary_iop!(RemAssign, %=, rem_assign);
-define_binary_iop!(BitAndAssign, &=, bitand_assign);
-define_binary_iop!(BitOrAssign, |=, bitor_assign);
-define_binary_iop!(ShlAssign, <<=, shl_assign);
-define_binary_iop!(ShrAssign, >>=, shr_assign);
